backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
redis==5.0.8
celery==5.4.0
PyJWT==2.9.0
email-validator==2.2.0
httpx==0.27.2
itsdangerous==2.2.0
boto3==1.35.24
minio==7.2.10
passlib[bcrypt]==1.7.4
